requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
aio-pika==9.5.5
prometheus-client==0.21.1
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation-fastapi==0.52b1

SQLAlchemy==2.0.36
asyncpg==0.29.0
greenlet==3.1.1
aiosqlite==0.20.0
redis==5.2.1

python-dotenv==1.0.1
